+++ src/features/auth/email-verification/ui/email-verification-example.stories.tsx
@@ -1,9 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/nextjs'
-import { FormProvider } from 'react-hook-form'
-import FormField from '@/features/auth/ui/form-field'
-import InputWithButton from '@/features/auth/ui/input-with-button'
-import { useEmailVerification } from '../model/use-email-verification'
-import { useEmailVerificationForm } from '../model/use-email-verification-form'
+import EmailVerificationForm from './email-verification-form'
 
 const meta: Meta = {
   title: 'Features/Auth/EmailVerification/Example',
@@ -14,84 +10,18 @@ export default meta
 type Story = StoryObj
 
 export const EmailVerification: Story = {
-  render: () => {
-    const form = useEmailVerificationForm()
-    const emailVerification = useEmailVerification({
-      onSendCode: async email => {
-        await new Promise(r => setTimeout(r, 500))
-        alert(`인증코드가 ${email}로 발송되었습니다`)
-      },
-      onVerifyCode: async code => {
-        await new Promise(r => setTimeout(r, 500))
-        return code === '123456'
-      },
-    })
-    return (
-      <div style={{ width: 320 }}>
-        <FormProvider {...form}>
-          <form>
-            <FormField name="email" control={form.control} label="Email">
-              {field => (
-                <InputWithButton
-                  inputProps={{
-                    ...field,
-                    type: 'email',
-                    placeholder: 'Enter a valid email',
-                    disabled:
-                      emailVerification.isSending ||
-                      emailVerification.isVerified,
-                  }}
-                  buttonProps={{
-                    type: 'button',
-                    onClick: () =>
-                      emailVerification.sendCode(form.getValues('email')),
-                    disabled:
-                      emailVerification.isSending ||
-                      emailVerification.isVerified,
-                  }}
-                  buttonText={
-                    emailVerification.isSending ? 'Sending...' : 'Send'
-                  }
-                />
-              )}
-            </FormField>
-            {emailVerification.isCodeSent && (
-              <FormField name="code" control={form.control} label="인증코드">
-                {field => (
-                  <InputWithButton
-                    inputProps={{
-                      ...field,
-                      placeholder: '코드 6자리',
-                      maxLength: 6,
-                      disabled:
-                        emailVerification.isVerifying ||
-                        emailVerification.isVerified,
-                    }}
-                    buttonProps={{
-                      type: 'button',
-                      onClick: () =>
-                        emailVerification.verifyCode(form.getValues('code')),
-                      disabled:
-                        emailVerification.isVerifying ||
-                        emailVerification.isVerified,
-                      variant: emailVerification.isVerified
-                        ? 'outline'
-                        : 'default',
-                    }}
-                    buttonText={
-                      emailVerification.isVerifying
-                        ? 'Verifying...'
-                        : emailVerification.isVerified
-                          ? 'Verified'
-                          : 'Confirm'
-                    }
-                  />
-                )}
-              </FormField>
-            )}
-          </form>
-        </FormProvider>
-      </div>
-    )
-  },
+  render: () => (
+    <div style={{ width: 320 }}>
+      <EmailVerificationForm
+        onSendCode={async email => {
+          await new Promise(r => setTimeout(r, 500))
+          alert(`인증코드가 ${email}로 발송되었습니다`)
+        }}
+        onVerifyCode={async code => {
+          await new Promise(r => setTimeout(r, 500))
+          return code === '123456'
+        }}
+      />
+    </div>
+  ),
 }
